WIP: ooc inner join
