+++ polars/polars-lazy/polars-pipe/src/executors/sinks/joins/mod.rs
@@ -1,5 +1,6 @@
 #[cfg(feature = "cross_join")]
 mod cross;
+mod inner;
 
 #[cfg(feature = "cross_join")]
 pub(crate) use cross::*;
