+++ polars/polars-lazy/polars-pipe/src/executors/sinks/mod.rs
@@ -1,6 +1,7 @@
 pub(crate) mod groupby;
 mod joins;
 mod ordered;
+mod utils;
 
 #[cfg(feature = "cross_join")]
 pub(crate) use joins::*;
